refactor(compiler): split lib.rs into focused submodules
Convert crates/zccache-compiler/src/lib.rs (2565 lines) into focused
submodules so every file stays under 1000 LOC. AI tools and reviewers
can now navigate the parser without loading the whole file. No
behavioral changes - purely a code split. Public API is unchanged
(re-exports preserve every previously-pub item).

Layout (new files):
- detect.rs: detect_family, is_clang_cl_name, is_source_file,
  SOURCE_EXTENSIONS, MODULE_EXTENSIONS
- parse.rs: default_output, parse_invocation,
  source_mode_from_language, source_mode_from_extension
- parse_rustc.rs: rustc filename helpers + parse_rustc_invocation
- tests/mod.rs: declares submodules + shared args() helper
- tests/cpp_parse.rs: clang/gcc -c, multi-file, -x header mode,
  header units, sticky-mode regressions
- tests/cpp_output.rs: default output, PCH naming, concatenated -o,
  unknown-flag preservation, BUG_LINKER repro
- tests/detect.rs: detect_family + supports_depfile for
  clang/gcc/msvc/emcc
- tests/rustc.rs: crate types, --emit, --out-dir, proc-macro/bin
  output naming
- tests/clippy_driver.rs: clippy-driver detection + caching
- tests/modules.rs: C++20 modules, header units, --precompile,
  GCC -fmodules-ts
- tests/clang_cl.rs: clang-cl/cl.exe dispatch into MSVC parser
  (issue #261), MSVC-style flags

lib.rs now contains only the public types (CompilerFamily,
ParsedInvocation, CacheableCompilation), pub(crate) SourceMode, module
declarations, and pub use re-exports of detect_family and
parse_invocation. The pre-existing pub mod declarations (arduino,
parse_archiver, parse_linker, parse_msvc, parse_rustfmt, response_file,
strict_paths) are untouched.

parse_linker.rs updated its one `super::source_mode_from_language` call
to `crate::parse::source_mode_from_language` since that helper now
lives in the private parse module (pub(crate) within the crate).

Per-file LOC after split (all < 1000, tests target <= 600):
  lib.rs                 143
  detect.rs               64
  parse.rs               298
  parse_rustc.rs         296
  tests/mod.rs            20
  tests/cpp_parse.rs     426
  tests/cpp_output.rs    239
  tests/detect.rs         57
  tests/rustc.rs         402
  tests/clippy_driver.rs  94
  tests/modules.rs       379
  tests/clang_cl.rs      228

All 317 zccache-compiler unit tests pass.
\`cargo clippy --workspace --all-targets -- -D warnings\` is clean.
\`RUSTDOCFLAGS=-D warnings cargo doc -p zccache-compiler --no-deps\` is clean.
